feat: Make `opts` optional for greater flexibility in plugin usage
This change allows the opts property to be optional, enabling simpler
plugin configurations such as:

```lua
{
  "WTFox/jellybeans.nvim",
  lazy = false,
  priority = 1000,
  config = function()
    vim.cmd.colorscheme("jellybeans")
  end,
}
```

This makes the plugin easier to use in a variety of contexts—such as
cloning it directly into the Neovim plugin directory—without requiring
additional configuration or package manager.

**Why this PR:**

I'm the author of [vimcolorschemes.com](https://vimcolorschemes.com),
and I noticed `jellybeans.nvim` wasn't appearing on the site. That's
because the site's preview generator installs colorschemes by cloning
them locally, without passing any options, meaning plugins requiring
`opts` are skipped.

Making `opts` optional resolves this and allows your colorscheme to be
featured on the site. I did it manually for now, here's [the
link](https://vimcolorschemes.com/wtfox/jellybeans.nvim).

This is just a suggestion, and I'm happy to adjust or allow you to
adjust the PR if you have any feedback or prefer a different approach.
Thanks!